fix: polyzone detection

+++ data/resources/tsv/src/server/libs/zone/zone.ts
@@ -1,8 +1,10 @@
-import { Vector3, Crypto, Color } from '../../../core/libs';
+import { Entity, Vector3, Crypto, Color } from '../../../core/libs';
 import { User } from '../../../core/libs/user';
 import { Bucket } from '../bucket';
 import { IUser } from '../../../core/declares/user';
 import { IZone, ZoneType } from '../../../core/declares/zone';
+import { PolyZoneType, EntityZoneType } from '../../../core/declares/zone/types/zone';
+import { tsv } from '../..';
 
 abstract class Zone implements IZone {
   id: string;
@@ -10,15 +12,14 @@ abstract class Zone implements IZone {
   module: string;
   points: Vector3[] | Vector3;
   center?: Vector3;
-  min: Vector3;
-  max: Vector3;
-  size: Vector3;
-  area: number;
-  useGrid: boolean;
-  lazyGrid: boolean;
-  gridDivisions: number;
-  debugPoly: boolean;
-  debugGrid: boolean;
+  min?: Vector3;
+  max?: Vector3;
+  size?: Vector3;
+  useGrid?: boolean;
+  lazyGrid?: boolean;
+  gridDivisions?: number;
+  debugPoly?: boolean;
+  debugGrid?: boolean;
   color: {
     outline: Color;
     wall: Color;
@@ -29,10 +30,6 @@ abstract class Zone implements IZone {
   onLeave?: (user: User) => void;
 
   constructor(zone: ZoneType) {
-    if (zone.points instanceof Array && zone.points.length < 3) {
-      throw new Error('Zone must have at least 3 points');
-    }
-
     this.id = Crypto.uuidv4();
     this.name = zone.name;
     this.points = zone.points;
@@ -53,25 +50,33 @@ abstract class Zone implements IZone {
         };
 
     // this.bucket = zone.bucket && zone.bucket;
+    this.users = [];
+    this.onEnter = (user: User) => {
+      this.bucket !== undefined && this.bucket.addUser(user);
 
-    if (zone.onEnter)
-      this.onEnter = (user: User) => {
-        if (this.bucket) {
-          this.bucket.addUser(user);
-        }
-        zone.onEnter(user);
-      };
-    if (zone.onLeave)
-      this.onLeave = (user: User) => {
-        if (this.bucket) {
-          this.bucket.removeUser(user);
-        }
-        zone.onLeave(user);
-      };
+      this.addOneUser(user);
+
+      console.log('trigger');
+      console.log(user.source);
+      tsv.events.trigger({
+        name: 'onEnter',
+        module: 'zone',
+        onNet: true,
+        target: user.source,
+        data: [zone],
+      });
+
+      zone.onEnter && zone.onEnter(user);
+    };
+
+    this.onLeave = (user: User) => {
+      this.bucket && this.bucket.removeUser(user);
+      zone.onLeave && zone.onLeave(user);
+    };
   }
 
   abstract isInside(coords: Vector3): boolean;
-  abstract left(...args: Array<Vector3>): boolean;
+  abstract isLeft(...args: Array<Vector3>): number;
 
   addOneUser(user: IUser): boolean {
     if (this.users.find((u) => u.id === user.id)) {
@@ -84,22 +89,27 @@ abstract class Zone implements IZone {
 }
 
 class PolyZone extends Zone {
-  private innerLoop(p0: Vector3, p1: Vector3, p2: Vector3, i: number): number {
+  polygon: Vector3[];
+  gridArea: number;
+  gridCellWidth: number;
+  gridCellHeigth: number;
+
+  private innerLoop(p0: Vector3, p1: Vector3, p2: Vector3, wn: number): number {
     if (p0.y <= p2.y) {
       if (p1.y > p2.y) {
-        if (this.left(p0, p1, p2)) {
-          return ++i;
+        if (this.isLeft(p0, p1, p2) > 0) {
+          return wn + 1;
         }
       }
     } else {
       if (p1.y <= p2.y) {
-        if (this.left(p0, p1, p2)) {
-          return --i;
+        if (this.isLeft(p0, p1, p2) < 0) {
+          return wn - 1;
         }
       }
     }
 
-    return i;
+    return wn;
   }
   private calculatePolygon() {
     if (
@@ -115,9 +125,7 @@ class PolyZone extends Zone {
         maxY = Number.MIN_SAFE_INTEGER,
         maxZ = Number.MIN_SAFE_INTEGER;
 
-      const points = this.points as Vector3[];
-
-      points.forEach((point) => {
+      this.polygon.forEach((point) => {
         minX = Math.min(minX, point.x);
         minY = Math.min(minY, point.y);
         minZ = Math.min(minZ, point.z);
@@ -133,8 +141,12 @@ class PolyZone extends Zone {
     }
   }
 
-  constructor(zone: ZoneType) {
-    super(zone);
+  constructor(zone: PolyZoneType) {
+    super({
+      ...zone,
+      points: zone.polygon,
+    });
+    this.polygon = zone.polygon;
     this.calculatePolygon();
   }
 
@@ -148,46 +160,56 @@ class PolyZone extends Zone {
     ) {
       return false;
     }
-
     // Checks if point is within the polygon's height bounds
-    if ((this.min.z && coords.z < this.min.z) || (this.max.z && coords.z > this.max.z)) {
+    if ((this.min.z && coords.z - 1 < this.min.z) || (this.max.z && coords.z - 1 > this.max.z)) {
       return false;
     }
 
     // Returns true if the grid cell associated with the point is entirely inside the poly
-    if (this.useGrid) {
-      const gridPosX = coords.x - this.min.x;
-      const gridPosY = coords.y - this.min.y;
-      const gridCellX = gridPosX * this.gridDivisions; // size.x
-      const gridCellY = gridPosY * this.gridDivisions; // size.y
-      //   const gridCellValue = grid[gridCellY + 1][gridCellX + 1]
+    // if (this.useGrid) {
+    //   const gridPosX = coords.x - this.min.x;
+    //   const gridPosY = coords.y - this.min.y;
+    //   const gridCellX = gridPosX * this.gridDivisions; // size.x
+    //   const gridCellY = gridPosY * this.gridDivisions; // size.y
+    //   //   const gridCellValue = grid[gridCellY + 1][gridCellX + 1]
 
-      //   if gridCellValue == nil and poly.lazyGrid then
-      //     gridCellValue = _isGridCellInsidePoly(gridCellX, gridCellY, poly)
-      //     grid[gridCellY + 1][gridCellX + 1] = gridCellValue
-      //   end
-      //   if gridCellValue then return true end
-      // end
+    //   //   if gridCellValue == nil and poly.lazyGrid then
+    //   //     gridCellValue = _isGridCellInsidePoly(gridCellX, gridCellY, poly)
+    //   //     grid[gridCellY + 1][gridCellX + 1] = gridCellValue
+    //   //   end
+    //   //   if gridCellValue then return true end
+    //   // end
+    // }
+
+    let wn = 0;
+
+    for (let i = 0; i < this.polygon.length - 1; i++) {
+      wn = this.innerLoop(this.polygon[i], this.polygon[i + 1], coords, wn);
     }
-
-    if (this.points instanceof Array) {
-      let wn = 0;
-      for (let j = 0; j < this.points.length; j++) {
-        wn = this.innerLoop(this.points[wn], this.points[++wn], coords, wn);
-      }
-      wn = this.innerLoop(this.points[wn], this.points[0], coords, wn);
-
-      return wn !== 0;
-    }
+    wn = this.innerLoop(this.polygon[this.polygon.length - 1], this.polygon[0], coords, wn);
+    return wn !== 0;
   }
-  left(p0: Vector3, p1: Vector3, p2: Vector3): boolean {
-    return (p1.x - p0.x) * (p2.y - p0.y) - (p2.x - p0.x) * (p1.y - p0.y) > 0;
+  isLeft(p0: Vector3, p1: Vector3, p2: Vector3): number {
+    return (p1.x - p0.x) * (p2.y - p0.y) - (p2.x - p0.x) * (p1.y - p0.y);
   }
 }
-// class EntityZone extends Zone {
-//   constructor(zone: EntityZoneType) {
-//     super(zone);
-//   }
-// }
+class EntityZone extends Zone {
+  entity: Entity;
 
-export { Zone, PolyZone /*EntityZone*/ };
+  constructor(zone: EntityZoneType) {
+    super({
+      ...zone,
+      points: zone.entity.Position,
+    });
+    this.entity = zone.entity;
+  }
+
+  isInside(coords: Vector3): boolean {
+    return false;
+  }
+  isLeft(p0: Vector3, p1: Vector3, p2: Vector3): number {
+    return 0;
+  }
+}
+
+export { Zone, PolyZone, EntityZone };
